glue Cull + refactor move
