+++ src/main/java/config/EvolutionConfig.java
@@ -1,4 +1,4 @@
-package logic;
+package config;
 
 /**
  * EvolutionConfig:
