+++ src/main/java/logic/GeneticAlgorithm.java
@@ -1,6 +1,9 @@
 package logic;
 
+import config.CullConfig;
+import config.EvolutionConfig;
 import data.GenePool;
+import model.GeneticCandidateSolution;
 
 /**
  * GeneticAlgorithm:
@@ -9,33 +12,32 @@ public class GeneticAlgorithm {
     private static final int max = 3000000;
 
     private final GenePool<GeneticCandidateSolution> genepool;
-    private final Object cullConfig;
-    //TODO require interface code pulled
-//    private final Cull cullImplementor;
+    private final CullConfig cullConfig;
+    private final EvolutionConfig config;
+    private Cullable cullImplementor;
 //    private final Combine combineImplementor;
     private final Object combineConfig;
-    private final EvolutionConfig evolutionConfig;
+    private final int killAmount = 5;
+    private final double probablityToKillGoodGuys = 0.2;
 
-    public GeneticAlgorithm() {
+    public GeneticAlgorithm(EvolutionConfig cfg) {
         this.genepool = new GenePool<GeneticCandidateSolution>();
-        // note that all these are dependencies, the following lines are demo
-        // only, don't copy paste
-        this.cullConfig = new Object(); // cullconfig injection
+
+        this.cullConfig = new CullConfig(killAmount, probablityToKillGoodGuys);
+        // cullconfig
+        // injection
         this.combineConfig = new Object();
-        this.evolutionConfig = new EvolutionConfig();
-        evolutionConfig.setStepMax(max);
+        this.config = cfg;
         //TODO pull request some of the interface first please
-//        this.cullImplementor = new Cull(cullConfig,genepool);
+        this.cullImplementor = new Cull();
 //        this.combineImplementor = new Combine(combineConfig, genepool);
 
     }
 
     public GeneticCandidateSolution templateSolve() {
         populationInitialSeeding();
-        for (int i = 0; i < evolutionConfig.evolutionStepMax; i++){
-            //TODO pull request some of the interface code please so that
-            // this can compile
-//            cullImplementor.cull(this.cullConfig, this.genepool);
+        for (int i = 0; i < config.evolutionStepMax; i++){
+            cullImplementor.cull(this.cullConfig, this.genepool);
 //            combineConfig.combine(this.combineConfig,this.genepool);
         }
         return this.genepool.getBest();
@@ -44,6 +46,6 @@ public class GeneticAlgorithm {
     // do some initial step to stem up the population. thrive my minions, i
     // grant you 3 years of time to snag without natural predator
     private void populationInitialSeeding() {
-
+        //TODO
     }
 }
