+++ src/main/java/model/GeneticCandidateSolution.java
@@ -1,11 +1,12 @@
-package logic;
+package model;
 
+import model.CandidateSolution;
 import model.DataSource;
 
 /**
  * GeneticCandidateSolution:
  */
-public class GeneticCandidateSolution extends CandidateSolution{
+public class GeneticCandidateSolution extends CandidateSolution {
     /**
      * @param prefs
      */
