+++ src/test/java/logic/ImprovableCandidateSolutionTest.java
@@ -2,6 +2,7 @@ package logic;
 
 import app.Config;
 import model.DataSource;
+import model.ImprovableCandidateSolution;
 import org.junit.Before;
 import org.junit.Test;
